requirements.txt
absl-py==1.2.0
batchgenerators==0.24
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==37.0.1
cycler==0.11.0
dicom2nifti==2.4.3
einops==0.6.0
fonttools==4.37.1
future==0.18.2
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grad-cam==1.4.6
grpcio==1.47.0
idna==3.3
imageio==2.21.0
joblib==1.1.0
kiwisolver==1.4.4
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==0.9.1
networkx==2.8.5
nibabel==4.0.1
nnunet==1.7.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pip==22.2.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-gdcm==3.0.14
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.1
setuptools==63.3.0
SimpleITK==2.1.1.2
six==1.16.0
sklearn==0.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.7.31
torch-tb-profiler==0.4.0
tqdm==4.64.0
traceback2==1.4.0
ttach==0.0.3
typing_extensions==4.3.0
unittest2==1.1.0
urllib3==1.26.11
Werkzeug==2.2.2
wheel==0.37.1
yacs==0.1.8
pyyaml==6.0